annotation fix

+++ python_async_comprehension/1-async_comprehension.py
@@ -6,6 +6,6 @@ from typing import List
 async_generator = __import__('0-async_generator').async_generator
 
 
-async def async_comprehension() -> List[int]:
+async def async_comprehension() -> List[float]:
     """Coroutine return the 10 random numbers"""
     return [i async for i in async_generator()]
